God exists! ;-)
This is the first version of the TPTP THF formalization for which LEO-II and Satallax are able to prove the last theorem.
The modifications with respect to the last version were made by Christoph Benzmüller on 11th of July 2013.
The modified files were received by Bruno Woltzenlogel Paleo by email on 11th of July 2013 at 17:09.